Enum mas_jose::jwa::AsymmetricSigningKey

#[non_exhaustive]
pub enum AsymmetricSigningKey {
    Rs256(Rs256SigningKey),
    Rs384(Rs384SigningKey),
    Rs512(Rs512SigningKey),
    Ps256(Ps256SigningKey),
    Ps384(Ps384SigningKey),
    Ps512(Ps512SigningKey),
    Es256(Es256SigningKey),
    Es384(Es384SigningKey),
    Es256K(Es256KSigningKey),
}

An enum of all supported asymmetric signature algorithms verifying keys

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Rs256(Rs256SigningKey)

Rs384(Rs384SigningKey)

Rs512(Rs512SigningKey)

Ps256(Ps256SigningKey)

Ps384(Ps384SigningKey)

Ps512(Ps512SigningKey)

Es256(Es256SigningKey)

Es384(Es384SigningKey)

Es256K(Es256KSigningKey)

Implementations

impl AsymmetricSigningKey

pub fn rs256(key: RsaPrivateKey) -> Self

Create a new signing key with the RS256 algorithm from the given RSA private key.

pub fn rs384(key: RsaPrivateKey) -> Self

Create a new signing key with the RS384 algorithm from the given RSA private key.

pub fn rs512(key: RsaPrivateKey) -> Self

Create a new signing key with the RS512 algorithm from the given RSA private key.

pub fn ps256(key: RsaPrivateKey) -> Self

Create a new signing key with the PS256 algorithm from the given RSA private key.

pub fn ps384(key: RsaPrivateKey) -> Self

Create a new signing key with the PS384 algorithm from the given RSA private key.

pub fn ps512(key: RsaPrivateKey) -> Self

Create a new signing key with the PS512 algorithm from the given RSA private key.

pub fn es256(key: SecretKey<NistP256>) -> Self

Create a new signing key with the ES256 algorithm from the given ECDSA private key.

pub fn es384(key: SecretKey<NistP384>) -> Self

Create a new signing key with the ES384 algorithm from the given ECDSA private key.

pub fn es256k(key: SecretKey<Secp256k1>) -> Self

Create a new signing key with the ES256K algorithm from the given ECDSA private key.

pub fn from_jwk_and_alg( params: &JsonWebKeyPrivateParameters, alg: &JsonWebSignatureAlg ) -> Result<Self, AsymmetricKeyFromJwkError>

Create a new signing key for the given algorithm from the given private JWK parameters.

Errors

Returns an error if the key parameters are not suitable for the given algorithm.